Remove stuck double spent pending txs where bdk assumes not relevant
If a user was outbid while their transaction is still in the mempool, the pending transaction gets stuck in the wallet which could result in building invalid chains of transactions on top.

The solution is either to rescan the wallet to forget the tx or, for now, we just watch bids in memory and manually insert them to the chain so that bdk can purge the double spent tx.

This is a temporary solution until bdk does this automatically or adds better tx invalidation APIs